This patch addresses the threading issues with the lexer. This requires flex 2.3.51 which can generate reentrant scanners. Now, instead of using a global variable, a yyscan_t structure is allocated and passed along to every call into flex.
Sterling Hughes will have a hard time recognizing his code after so much
of my tweaking it :-)

The scanner definition was changed to send as much as possible of HTML in
one triple-quoted string, rather than line-by-line.

The copyright in psp_parser.l is fixed, also added a little blurb explaining
what's going on there.

I had to once again revert from "print" to "req.write" because we can't touch
sys.stdout because it is shared among threads.

The caching function is disabled for now until we make it thread-safe.

A new flex-generated header file psp_flex.h was added.

Finally - just a warning - PSP still needs a lot of work, this is only the
first step of many more coming.