The PythonImport directive makes use of the handler list data structure even though it specifically is for a different purpose. The handler list data structure needs to be changed a bit in order to support changes required to address MODPYTHON-63 and MODPYTHON-126. These changes may make it unusable in the way that the PythonImport directive was using it, or at least it will not be a good match. The PythonImport directive has therefore been modified to use a standard Apache table object thereby allowing arbitrary changes to the handler list data structure to be made as necessary.
Note that although the current Apache table implementation probably results
in module imports occuring in the order they appear in the Apache
configuration file, if how it is implemented is changed, this may no longer
be the case. This is not seen as an issue as the documentation for the
PythonImport directive does not give any guarantees about ordering of
imports within the context of a specific interpreter and it would be bad
practice for modules to rely on mod_python importing them in a specific
order. If there are ordering requirements, it should be dealt with internal
to the module being imported so that it is self contained.